requirements.txt
fastapi>=0.118.0
python-multipart>=0.0.20
uvicorn>=0.37.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
